pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "gibson-cli"
version = "1.0.0"
description = "AI/ML Security Testing Framework - Developer-first CLI for comprehensive AI security testing"
authors = ["Zero-Day.ai <security@zero-day.ai>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/zero-day-ai/gibson-framework"
repository = "https://github.com/zero-day-ai/gibson-framework"
documentation = "https://github.com/zer-day-ai/gibson-framework/docs"
keywords = ["security", "ai", "ml", "llm", "testing", "cli", "owasp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]
packages = [{include = "gibson"}]
include = [
    "gibson/py.typed",
    "configs/*.yaml",
    "gibson/data/**/*",
]
exclude = [
    "tests/**/*",
    "docs/**/*",
    "*.pyc",
    "__pycache__",
]

[tool.poetry.scripts]
gibson = "gibson.main:main"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/zero-day-ai/gibson-cli/issues"
"Changelog" = "https://github.com/zero-day-ai/gibson-cli/blob/main/CHANGELOG.md"
"Documentation" = "https://docs.zero-day.ai/gibson"
"Source Code" = "https://github.com/zero-day-ai/gibson-cli"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
httpx = "^0.27.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
aiosqlite = "^0.19.0"
alembic = "^1.13.0"
pyyaml = "^6.0.1"
gitpython = "^3.1.40"
keyring = "^24.3.0"
cryptography = "^41.0.7"
tenacity = "^8.2.3"
platformdirs = "^4.1.0"
validators = "^0.22.0"
prompt-toolkit = "^3.0.43"
aiofiles = "^23.2.1"
psutil = "^5.9.6"
orjson = "^3.9.10"
loguru = "^0.7.2"
python-dateutil = "^2.8.2"
aiocache = "^0.12.2"
asyncclick = "^8.1.7"
ratelimit = "^2.2.1"
aiohttp = "^3.12.15"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.27.1"
litellm = "^1.55.0"
jsonschema = "^4.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
pytest-timeout = "^2.2.0"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"
hypothesis = "^6.92.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
ipython = "^8.18.1"
nuitka = "^1.9.0"
respx = "^0.21.0"  # For mocking httpx requests in LiteLLM tests
faker = "^30.0.0"  # For generating test data

[tool.poetry.group.ai]
optional = true

[tool.poetry.group.ai.dependencies]
transformers = "^4.36.0"
torch = "^2.1.0"
langchain = "^0.0.350"
openai = "^1.6.0"
anthropic = "^0.8.0"

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "B904"]
fixable = ["ALL"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config --asyncio-mode=auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filerglob = "test_*.py"
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests that may require external services",
    "slow: Tests that take a long time to run",
    "asyncio: Async tests using pytest-asyncio",
    "database: Tests that require database setup",
    "cli: Tests that exercise CLI commands",
    "mock: Tests that use extensive mocking",
]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["gibson"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/*_test.py",
    "gibson/__main__.py",
    "gibson/main.py",
]
branch = true
concurrency = ["thread", "multiprocessing"]
data_file = ".coverage"
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "class .*\\(Protocol\\):",
    "except ImportError:",
]
show_missing = true
skip_covered = false
fail_under = 85
precision = 2
sort = "Cover"

[tool.coverage.html]
directory = "htmlcov"
title = "Gibson Framework Coverage Report"

[tool.coverage.xml]
output = "coverage.xml"
